Extracted canvas from our page as a component; fixed broadcasting issues on real-time editting(including add box, change content, delete box
