+++ client/src/app/draw/draw.css
@@ -1,0 +1,29 @@
+.canvas-container {
+    background-color: var(--background-color);
+    padding: 30px 25px;
+    position: relative;
+  }
+  
+  .drawing-canvas {
+    width: 674px;
+    height: 953px;
+    border: 1px solid #ccc;
+    background-color: white;
+    display: block;
+    margin: 0 auto;
+  }
+  
+  .canvas-textbox {
+    position: absolute;
+    font-size: 16px;
+    line-height: 16px;
+    padding: 2px;
+    background: white;
+    z-index: 15;
+    border: 1px solid transparent;
+  }
+  
+  .canvas-textbox.editing {
+    border: 1px solid blue;
+  }
+
